Finished very to module (LabB)
I imported the switches, hex, etc. .txt file
I connected the processor to the hex displays.
Were pretty much ready to compile. However, we still need
to implement a few changes in the settings, such as changing the
settings so that quartus recognizes the state machine, change the
settings so that we can read and write to the ROM and RAM while it's
running on DE2(Look at Homework 6), modify the .mif files according to
the instructions. We should meet up tomorrow after class and finish it
up, that way we can get started on the lab report.